==> 07-create-training-data.ipynb <==
import pandas as pd
import numpy as np

# get all stars df
allstars = pd.read_csv('./final_data/allstars.csv')
# AS column to indentify if a player was an all star in respective season
allstars['AS'] = ''
allstars = allstars.rename(columns={'Player': 'PLAYER_NAME', 'Years' : 'SEASON'})
allstars['SEASON'] = allstars['SEASON'].apply(lambda x: str(x))
allstars        

# get final player stats df
stats = pd.read_csv('./final_data/final_player_stats.csv')
# convert season to the year that the all star game took place in the season (1996-97 --> 1997)
stats['SEASON'] = stats['SEASON'].str[:2] + stats['SEASON'].str[5:7]
stats['SEASON'] = stats['SEASON'].replace('1900', '2000')
stats

# merge allstars with player stats to get only all stars, make AS == 1 for all rows
AS_stats = pd.merge(allstars, stats, on = ['PLAYER_NAME', 'SEASON'])
AS_stats['AS'] = 1
AS_stats

# join dataframes together, drop duplicates, replace NaN in AS column with 0
final_df = pd.concat([stats, AS_stats], ignore_index=True)
final_df = final_df.drop_duplicates(subset=['PLAYER_NAME', 'SEASON'], keep='last')
final_df['AS'] = final_df['AS'].fillna(0, inplace=False)
final_df

# save as CSV
final_df.to_csv('./final_data/final_data.csv')
